webapp/requirements.txt
fastapi
uvicorn[standard]    
python-multipart     
ollama               
openai-whisper       
numpy
scipy                
python-dotenv
aiofiles             